Easier testing
- new scripts ./hack/run-test.sh and ./hack/run-e2e-test.sh

- Easy e2e testing: set env vars to skip installation and teardown

- Makefile runs 'docker buildx build' only when it must

- Makefile fix for LC_DYSYMTAB error on MacOS

- TestRenderHelmChart mocks the kgw version

+++ test/kubernetes/e2e/test.go
@@ -135,6 +135,9 @@ func (i *TestInstallation) InstallRevisionedIstio(ctx context.Context, rev, prof
 }
 
 func (i *TestInstallation) UninstallIstio() error {
+	if testutils.ShouldSkipIstioInstall() || testutils.ShouldSkipTeardown() {
+		return nil
+	}
 	return cluster.UninstallIstio(i.IstioctlBinary, i.ClusterContext.KubeContext)
 }
 
@@ -150,6 +153,13 @@ func (i *TestInstallation) InstallKgatewayFromLocalChart(ctx context.Context) {
 func (i *TestInstallation) InstallKgatewayCRDsFromLocalChart(ctx context.Context) {
 	if testutils.ShouldSkipInstall() {
 		return
+	}
+
+	// Check if we should skip installation if the release already exists
+	if testutils.ShouldSkipInstallIfPresent() {
+		if i.Actions.Helm().ReleaseExists(ctx, helmutils.CRDChartName, i.Metadata.InstallNamespace) {
+			return
+		}
 	}
 
 	// install the CRD chart first
@@ -169,6 +179,13 @@ func (i *TestInstallation) InstallKgatewayCRDsFromLocalChart(ctx context.Context
 func (i *TestInstallation) InstallKgatewayCoreFromLocalChart(ctx context.Context) {
 	if testutils.ShouldSkipInstall() {
 		return
+	}
+
+	// Check if we should skip installation if the release already exists
+	if testutils.ShouldSkipInstallIfPresent() {
+		if i.Actions.Helm().ReleaseExists(ctx, helmutils.ChartName, i.Metadata.InstallNamespace) {
+			return
+		}
 	}
 
 	// and then install the main chart
@@ -201,7 +218,7 @@ func (i *TestInstallation) UninstallKgateway(ctx context.Context) {
 }
 
 func (i *TestInstallation) UninstallKgatewayCore(ctx context.Context) {
-	if testutils.ShouldSkipInstall() {
+	if testutils.ShouldSkipInstall() || testutils.ShouldSkipTeardown() {
 		return
 	}
 
@@ -218,7 +235,7 @@ func (i *TestInstallation) UninstallKgatewayCore(ctx context.Context) {
 }
 
 func (i *TestInstallation) UninstallKgatewayCRDs(ctx context.Context) {
-	if testutils.ShouldSkipInstall() {
+	if testutils.ShouldSkipInstall() || testutils.ShouldSkipTeardown() {
 		return
 	}
